evaluators: enable multiline commands
* enable botlog for .term while at it
* refactor security checks

the check doesn't even work most of the time; it can be easily
bypassed. let's use find() instead of "in" method to check for dangerous
commands. this has the potential for false positives so consider this
commit as wip. i'll patch false positives when i get reports.

let's also patch ".term echo $ENV_VARS" and variations of it.

Signed-off-by: KenHV <kenhv@null.net>
Signed-off-by: BianSepang <bianconneri46@gmail.com>